Refactor to make a common web UI entry point
We will have a few different places to setup the web UI (origin, namespace,
director); start consolidating them.
