+++ cmd/origin_serve.go
@@ -12,7 +12,6 @@ import (
 	"os"
 	"os/exec"
 	"os/signal"
-	"net/http"
 	"path"
 	"path/filepath"
 	"strings"
@@ -20,13 +19,13 @@ import (
 	"text/template"
 	"time"
 
-	"github.com/gin-gonic/gin"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"github.com/pelicanplatform/pelican"
 	"github.com/pelicanplatform/pelican/config"
 	"github.com/pelicanplatform/pelican/origin_ui"
+	"github.com/pelicanplatform/pelican/web_ui"
 	"github.com/pkg/errors"
 )
 
@@ -510,44 +509,21 @@ func serve(/*cmd*/ *cobra.Command, /*args*/ []string) error {
 		return err
 	}
 
-	gin.SetMode(gin.ReleaseMode)
-	engine := gin.New()
-	engine.Use(gin.Recovery())
-	webLogger := log.WithFields(log.Fields{"daemon": "gin"})
-	engine.Use(func(ctx *gin.Context) {
-		startTime := time.Now()
-
-		ctx.Next()
-
-		latency := time.Since(startTime)
-		webLogger.WithFields(log.Fields{"method": ctx.Request.Method,
-			"status": ctx.Writer.Status(),
-			"time": latency.String(),
-			"client": ctx.RemoteIP(),
-			"resource": ctx.Request.URL.Path},
-		).Info("Served Request")
-	})
-	if err = pelican.ConfigureMetrics(engine); err != nil {
+	engine, err := web_ui.GetEngine()
+	if err != nil {
 		return err
 	}
 	if err = origin_ui.ConfigureOriginUI(engine); err != nil {
 		return err
 	}
 
-	engine.GET("/api/v1.0/health", func(c *gin.Context) {
-		c.JSON(http.StatusOK, gin.H{"status": "healthy"})
-	})
-
-	go func() {
-		err = engine.Run()
-		if err != nil {
-			panic(err)
-		}
-	}()
+	go web_ui.RunEngine(engine)
 
 	// Ensure we wait until the origin has been initialized
 	// before launching XRootD.
-	origin_ui.WaitUntilLogin()
+	if err = origin_ui.WaitUntilLogin(); err != nil {
+		return err
+	}
 
 	err = launchXrootd()
 	if err != nil {
