+++ origin_ui/origin.go
@@ -9,10 +9,12 @@ import (
 	"net/http"
 	"net/url"
 	"os"
+	"os/signal"
 	"path"
 	"path/filepath"
 	"strings"
 	"sync/atomic"
+	"syscall"
 	"time"
 
 	"github.com/gin-gonic/gin"
@@ -73,10 +75,13 @@ func periodicReload() {
 	}
 }
 
-func WaitUntilLogin() {
+func WaitUntilLogin() error {
 	if authDB.Load() != nil {
-		return
+		return nil
 	}
+	sigs := make(chan os.Signal, 1)
+	signal.Notify(sigs, syscall.SIGINT, syscall.SIGTERM)
+
 	for {
 		previousCode.Store(currentCode.Load())
 		newCode := fmt.Sprintf("%06v", rand.Intn(1000000))
@@ -85,9 +90,14 @@ func WaitUntilLogin() {
 		fmt.Println(*currentCode.Load())
 		start := time.Now()
 		for time.Since(start) < 30 * time.Second {
-			time.Sleep(100 * time.Millisecond)
+			select {
+			case <-sigs:
+				return errors.New("Process terminated...")
+			default:
+				time.Sleep(100 * time.Millisecond)
+			}
 			if authDB.Load() != nil {
-				return
+				return nil
 			}
 		}
 	}
