+++ web_ui/ui.go
@@ -1,0 +1,52 @@
+
+package web_ui
+
+import (
+	"net/http"
+	"time"
+
+	"github.com/gin-gonic/gin"
+	log "github.com/sirupsen/logrus"
+	"github.com/zsais/go-gin-prometheus"
+)
+
+func ConfigureMetrics(engine *gin.Engine) error {
+	prometheusMonitor := ginprometheus.NewPrometheus("gin")
+	prometheusMonitor.Use(engine)
+
+	engine.GET("/api/v1.0/health", func(ctx *gin.Context) {
+		ctx.JSON(http.StatusOK, gin.H{"status": "healthy"})
+	})
+	return nil
+}
+
+func GetEngine() (*gin.Engine, error) {
+	gin.SetMode(gin.ReleaseMode)
+	engine := gin.New()
+	engine.Use(gin.Recovery())
+	webLogger := log.WithFields(log.Fields{"daemon": "gin"})
+	engine.Use(func(ctx *gin.Context) {
+		startTime := time.Now()
+
+		ctx.Next()
+
+		latency := time.Since(startTime)
+		webLogger.WithFields(log.Fields{"method": ctx.Request.Method,
+			"status": ctx.Writer.Status(),
+			"time": latency.String(),
+			"client": ctx.RemoteIP(),
+			"resource": ctx.Request.URL.Path},
+		).Info("Served Request")
+	})
+	if err := ConfigureMetrics(engine); err != nil {
+		return nil, err
+	}
+	return engine, nil
+}
+
+func RunEngine(engine *gin.Engine) {
+	err := engine.Run()
+	if err != nil {
+		panic(err)
+	}
+}
